Create opinfo_utils
